+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Dropped unused `neusuw.php` file
 - Dropped unused field `users.sign`
 - Dropped unused field `users.kiler`
+- Dropped unused fields `users.kolorplus`, `users.kolorminus`, `users.kolorpoziom`
 - Dropped unused field `users.raids`
 - Dropped unused field `users.p_infligees`